+++ UE4Config.Tests/SanityTests.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.IO;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 
 namespace UE4Config.Tests
 {
